rename plugin to orchestrator, rename orchestrator to register
