+++ modules/api/src/main/java/org/eclipse/xpanse/api/OrchestratorApi.java
@@ -38,8 +38,8 @@ import org.eclipse.xpanse.modules.models.service.view.ServiceVo;
 import org.eclipse.xpanse.modules.models.service.view.UserAvailableServiceVo;
 import org.eclipse.xpanse.modules.plugin.deployment.DeployTask;
 import org.eclipse.xpanse.modules.plugin.deployment.Deployment;
-import org.eclipse.xpanse.orchestrator.OrchestratorService;
-import org.eclipse.xpanse.orchestrator.register.RegisterService;
+import org.eclipse.xpanse.modules.register.OrchestratorService;
+import org.eclipse.xpanse.modules.register.register.RegisterService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.hateoas.Link;
 import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
