+++ modules/orchestrator/src/main/java/org/eclipse/xpanse/modules/plugin/credential/AuthenticationCapabilities.java
