+++ modules/orchestrator/src/main/java/org/eclipse/xpanse/modules/plugin/utils/OpenApiUtil.java
