+++ modules/register/src/test/resources/ocl_testOpenApi.yaml
